Switch graphql-source for source-shopify2
Not using gatsby-source-shopify which is the official one as there seems to be errors on Gatsby v2. Should be a seamless switch in gatsby-plugins.js once it works again.

We're switching from the generic graphql-source to a Shopify specific source plugin in order to get the benefits of pagination and imageSharp. This does however have potential issues for larger stores as it retrieves ALL products, ALL collections, ALL images, etc. The graphql-source was easier to limit.

Due to how the graphql is structured, there is a considerable number of query structure changes